[Cache] Fix test failure related to Redis6Proxy on appveyor
