+++ src/Symfony/Component/Cache/Traits/Redis6Proxy.php
@@ -28,7 +28,6 @@ class Redis6Proxy extends \Redis implements ResetInterface, LazyObjectInterface
     use LazyProxyTrait {
         resetLazyObject as reset;
     }
-    use Redis6ProxyTrait;
 
     private const LAZY_OBJECT_PROPERTY_SCOPES = [];
 
@@ -225,6 +224,11 @@ class Redis6Proxy extends \Redis implements ResetInterface, LazyObjectInterface
     public function discard(): \Redis|bool
     {
         return ($this->lazyObjectState->realInstance ??= ($this->lazyObjectState->initializer)())->discard(...\func_get_args());
+    }
+
+    public function dump($key): \Redis|string
+    {
+        return ($this->lazyObjectState->realInstance ??= ($this->lazyObjectState->initializer)())->dump(...\func_get_args());
     }
 
     public function echo($str): \Redis|false|string
@@ -645,6 +649,11 @@ class Redis6Proxy extends \Redis implements ResetInterface, LazyObjectInterface
     public function ltrim($key, $start, $end): \Redis|bool
     {
         return ($this->lazyObjectState->realInstance ??= ($this->lazyObjectState->initializer)())->ltrim(...\func_get_args());
+    }
+
+    public function mget($keys): \Redis|array
+    {
+        return ($this->lazyObjectState->realInstance ??= ($this->lazyObjectState->initializer)())->mget(...\func_get_args());
     }
 
     public function migrate($host, $port, $key, $dstdb, $timeout, $copy = false, $replace = false, #[\SensitiveParameter] $credentials = null): \Redis|bool
